requirements.txt
google-generativeai
pdfplumber
python-docx
python-dotenv
pydantic